requirements.txt
PyGithub==1.59.1
qrcode==7.4.2
pillow==10.3.0
pyzbar==0.1.9
requests==2.31.0
urllib3==1.26.18
pyOpenSSL==23.2.0
cryptography==41.0.7
python-dotenv==1.0.1